unit tests [stage-1]
