+++ test/unit/template-editor/components/directives/dtv-component-image.spec.js
@@ -29,6 +29,9 @@ describe('directive: TemplateComponentImage', function() {
     $provide.service('templateEditorFactory', function() {
       return {};
     });
+    $provide.service('storageManagerFactory', function() {
+      return {};
+    });    
     $provide.service('componentsFactory', function() {
       return {
         selected: { id: "TEST-ID" },
